test: step2 JDBC 테스트 추가
